Session 5: Completed first draft of preprocessing
Only waves 1 and 2 are currently fetchable. Waves 3 to 5 have some changes I need to account for manually. I have moved the wave files to their own folder.

I have not included health features as of yet due to the nature of the dataset for section H. Waves 1 and 2 are exported to the CSV folder.

The model generator notebook combines the waves. I modified the LR model to rather take in a dataframe and target column.