Changes
Waves 1 and 2 preprocessing is now complete. Perhaps a few features can be revisited or new ones can be added but in total this should closely be final.

Implementations for all the models are also now in place. Each one uses similar approaches for cross validation and they all have similar structures class-wise.

I've also adjusted the preprocessing notebook so that now a bash script can be used to run the notebook with different parameters each time. The output is actually the ran notebook so a new nb is created for each wave.

Other than that we almost done really. But waves 3-5 certainly still need to be sorted out.